Modulos de socios, cuentas y agencias modificados
- Agencias
	- Se agrego un caso de uso para traer todas las agencias
	- Se agrego un controlador para procesar el request y traer todas las agencias
	- Se agrego una nueva ruta para traer todas las agencias
	- Se documento el endpoint de get para todas las agencias en SWAGGER
- Middleware
	- Para agencias se modifico la validacion de cantidad de ganadores para el ingreso
	- Para socios se agrego un middleware que verifique si la cuenta de un socio esta activa
- Se creo el controlador mensajes
	- Se implemento el envio de mensajes por SMS
	- Se comento el envio de mensajes, EN EVALUACION...
- Se agregaron los prefijos de actualizacion e eliminacion en cascada en la capa del modelo
- Recuperacion
	- Se agregaron los casos de uso para ingresar datos de recuperacion
		*Modificar datos de recuperacion
		*Generar codigos de recuperacion
		*Buscar los datos de una recuperacion
- Socios
	- Se agrego un caso de uso, controlador y ruta para buscar todos los socios de la tabla
		* Se documento el endpoint usando swagger
	- Se agrego un caso de uso, controlador y ruta para buscar un socio por el celular
		* Se documento el enpoint usando swagger
	- Se corrigio el proceso de actualizacion de la imagen de los socios
- Cuenta
	- Se agrego un caso de uso, controlador y ruta para buscar una cuenta por el id de la cuenta
		* Se documento el endpoint en swagger
	- Se redefinieron nombres en el proceso de buscar cuenta, por buscar cuenta por socio

+++ app/controllers/socios.controller.js
@@ -4,36 +4,31 @@ import {
     ingresarSocio,
     buscarSocio,
     actualizarSocio,
-    eliminarSocio
+    eliminarSocio,
+    buscarSocioByPhone,
+    buscarAllSocios
 } from "../use-cases/socios/index.js";
 import fs from 'fs';
 
 
 const ingresarSocioController = (req, res) => {
-
     function searchAgencia(id) {
-        return new Promise((res, rej) => {
+        return new Promise((resolve, rej) => {
             const buscar = buscarAgencia(id);
-            res(buscar);
+            resolve(buscar);
         });
     }
 
     function ingresar(socio) {
-
-        return new Promise((res, rej) => {
+        return new Promise((resolve, rej) => {
             const added = ingresarSocio({ ...socio }, 
                 req.headers['x-real-ip'] || req.connection.remoteAddress);
-            res(added);
+            resolve(added);
         });
     }
-    
+
     searchAgencia(req.body.idAgencia)
     .then(agencia => {
-        agencia.status !== 200 && 
-        res.status(agencia.status).send({
-            message: agencia.message
-        });
-
         return agencia.message.id;
     })
     .then(id => {
@@ -41,20 +36,16 @@ const ingresarSocioController = (req, res) => {
         return ingresar(req.body);
     })
     .then(result=>{
-
-        res.status(result.status).send({
+        return res.status(result.status).send({
             message: result.message,
         })
     })
     .catch(err=>{
-        
-        const __dirname = process.cwd();
-
-        console.log(err);
-
-        fs.unlinkSync( __dirname+err.message.errors[0].instance.imagen);
-
-        res.status(err.status).send({
+        if(err.message.errors){
+            const __dirname = process.cwd();
+            fs.unlinkSync( __dirname+err.message.errors[0].instance.imagen);
+        }
+        return res.status(err.status).send({
             message: err.message
         })
     });
@@ -82,21 +73,17 @@ const actualizarSocioController = (req, res) => {
         });
     }
     searchSocio(idSocio).then(socio => {
-        socio.status !== 200 &&
-        res.status(socio.status).send({
-            message: socio.message
-        });
-        
         return socio.message;
     })
     .then(datos => updateSocio(datos.id, datos.imagen, req.body))
     .then(result=>{
-        res.status(result.status).send({
+        return res.status(result.status).send({
             message: result.message,
         })
     })
     .catch(err=>{
-        res.status(err.status).send({
+        console.log(err);
+        return res.status(err.status).send({
             message: err.message
         })
     });
@@ -117,6 +104,11 @@ const buscarSocioController =  (req, res) => {
 const eliminarSocioController = (req, res) => {
     const { idSocio } = req.params;
 
+    if(!idSocio)
+        res.status(400).send({
+            message: "Ha ocurrido un error"
+        });
+
     function search (id){
         return new Promise((res, rej)=>{
             const buscar = buscarSocio(id);
@@ -132,10 +124,11 @@ const eliminarSocioController = (req, res) => {
     }
 
     search(idSocio).then(socio=>{
-        if(socio.status !== 200)
+        if(socio.status !== 200){
             res.status(socio.status).send({
                 message: socio.message
             });
+        }
         return socio.message.id;
     })
     .then(id => eliminar(id))
@@ -151,9 +144,50 @@ const eliminarSocioController = (req, res) => {
     });    
 }
 
+function existSocioByPhoneController (req, res) {
+    const { number } = req.params;
+
+    const search = buscarSocioByPhone(number);
+
+    search.then(resp => {
+        if(resp.status == 200){
+            res.status(200).send({
+                message: {
+                    existe: true,
+                    idSocio: resp.message.id,
+                }
+            });
+        }else{
+            res.status(resp.status).send({
+                message: resp.message
+            });
+        }
+    }).catch(err=>{
+        console.log(err);
+    })
+}
+
+function buscarAllSociosController (req, res) {
+
+    const buscar = buscarAllSocios();
+
+    buscar.then(socios=>{
+        if(socios.status === 200)
+            res.status(socios.status).send({
+                message: socios.message
+            });
+    }).catch(err=>{
+        res.status(err.status).send({
+            message: err.message
+        });
+    });    
+}
+
 export default Object.freeze({
     ingresarSocio: ingresarSocioController,
     buscarSocio: buscarSocioController,
     actualizarSocio: actualizarSocioController,
-    eliminarSocio: eliminarSocioController
+    eliminarSocio: eliminarSocioController,
+    existSocioByPhone: existSocioByPhoneController,
+    buscarAllSocios: buscarAllSociosController,
 });
